feat: user roles

+++ to-do/src/main/java/com/todolist/todo/services/UserService.java
@@ -7,6 +7,7 @@ import java.util.List;
 
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.BeanUtils;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
@@ -15,9 +16,14 @@ import com.todolist.todo.dtos.GetUserResponseDto;
 import com.todolist.todo.dtos.UpdateUserRequestDto;
 import com.todolist.todo.exceptions.BadRequestException;
 import com.todolist.todo.models.UserModel;
+import com.todolist.todo.models.UserModel.UserRole;
 import com.todolist.todo.repositories.UserRepository;
 
 import lombok.RequiredArgsConstructor;
+
+import com.auth0.jwt.JWT;
+import com.auth0.jwt.algorithms.Algorithm;
+import com.auth0.jwt.interfaces.DecodedJWT;
 
 @Service
 @RequiredArgsConstructor
@@ -25,6 +31,21 @@ public class UserService {
   private final UserRepository userRepository;
   private final BCryptPasswordEncoder passwordEncoder;
   private final ModelMapper modelMapper = new ModelMapper();
+  @Value("${api.security.token.secret}")
+  private String secretKey;
+
+  private UserModel getUserFromToken(String token) {
+
+    DecodedJWT decodedJWT = JWT.require(Algorithm.HMAC256(secretKey))
+        .build()
+        .verify(token.replace("Bearer ", ""));
+
+    String userId = decodedJWT.getSubject();
+
+    Optional<UserModel> user = userRepository.findById(UUID.fromString(userId));
+
+    return user.get();
+  }
 
   public GetUserResponseDto createUser(CreateUserRequestDto body) {
 
@@ -42,6 +63,10 @@ public class UserService {
 
     BeanUtils.copyProperties(body, user, "password");
 
+    if (body.role() == null) {
+      user.setRole(UserRole.USER);
+    }
+
     user.setPassword(passwordEncoder.encode(body.password()));
 
     userRepository.save(user);
@@ -58,17 +83,22 @@ public class UserService {
         .collect(Collectors.toList());
   }
 
-  public GetUserResponseDto getUserById(UUID userId) {
+  public GetUserResponseDto getUserById(UUID userId, String token) {
 
     Optional<UserModel> foundUser = userRepository.findById(userId);
     if (foundUser.isEmpty()) {
       throw new BadRequestException("This user does not exist");
     }
 
+    UserModel userLogged = this.getUserFromToken(token);
+    if (foundUser.get() != userLogged && userLogged.getRole() == UserRole.USER) {
+      throw new BadRequestException("You can only access your own profile");
+    }
+
     return modelMapper.map(foundUser.get(), GetUserResponseDto.class);
   }
 
-  public GetUserResponseDto updateUser(UUID userId, UpdateUserRequestDto body) {
+  public GetUserResponseDto updateUser(UUID userId, UpdateUserRequestDto body, String token) {
 
     Optional<UserModel> foundUser = userRepository.findById(userId);
     var user = foundUser.get();
@@ -77,7 +107,14 @@ public class UserService {
       throw new BadRequestException("This user does not exist!");
     }
 
-    BeanUtils.copyProperties(body, user);
+    UserModel userLogged = this.getUserFromToken(token);
+    if (foundUser.get() != userLogged && userLogged.getRole() == UserRole.USER) {
+      throw new BadRequestException("You can only update your own profile");
+    }
+
+    if (body.name() != null) {
+      user.setName(body.name());
+    }
 
     if (body.email() != null) {
       Optional<UserModel> userByEmail = userRepository.findByEmail(body.email());
@@ -95,6 +132,10 @@ public class UserService {
       user.setUsername(body.username());
     }
 
+    if (body.role() != null) {
+      user.setRole(body.role());
+    }
+
     if (body.password() != null) {
       user.setPassword(passwordEncoder.encode(body.password()));
     }
@@ -104,11 +145,17 @@ public class UserService {
     return modelMapper.map(user, GetUserResponseDto.class);
   }
 
-  public void deleteUser(UUID id) {
+  public void deleteUser(UUID id, String token) {
     Optional<UserModel> foundUser = userRepository.findById(id);
     if (foundUser.isEmpty()) {
       throw new BadRequestException("This user does not exist");
     }
+
+    UserModel userLogged = this.getUserFromToken(token);
+    if (foundUser.get() != userLogged && userLogged.getRole() == UserRole.USER) {
+      throw new BadRequestException("You can only delete your own profile");
+    }
+
     userRepository.delete(foundUser.get());
   }
 }
